Pass 'computerChoice' as an argument to 'playRound'
This makes it easier to add the string corresponding to
'computerChoice' to the UI: I'd otherwise have to return
'computerChoice' as part of a set of multiple values from playRound,
which makes the code more difficult and awkward than it has to be.

Doing this also happens to conform with the assignment's initial
instructions.