Move all resetting code into one place
At first, I was making a distinction between "internal reset", and
reset that involved managing the presence or absence of the resetting
button itself. While this might've been useful when developing this
feature, it's no longer needed, and furthermore could strike readers
of the code as an anti-pattern.

Also, I've added some comments in the surrounding code.